move contest log transform functions
